Next up: request parsing!
Similarly to our last step where we introduced the formatting of HTTP
responses, this step requires us to understand the formatting of HTTP
requests.

We build a (very naive) parser that assumes that the browser has sent us
a valid HTTP request and we break it down into its component parts. We
store those in hash which we'll call a request, and we'll differentiate
that from our previous "request" variable by renaming that one to
"http_request". This allows us to differentiate between the raw string
that we received, and the structured hash that we've turned it into.

We're really starting to develop a web server now! Try visiting a couple
of different urls on this server like http://127.0.0.1/whatever and you
will see it reflected in the message that gets printed to the browser.

Next up, we will want to start "doing something" based on which path was
requested.