Introduce a web application
At this point its worth our time to begin talking about the difference
between a "web server" and a "web application". For our purposes, we'll
consider the "web server" the part that's responsible for receiving HTTP
and responding with it, while the "web application" will think in terms
of request and response objects, and can introduce its own logic for
routing those requests.

We'll start quite simple, our WebApplication has a process route which
the WebServer will call with a request. Within the WebApplication, we
can do whatever we want with that request, so long as we return a
special response format, which we'll consider to be an array consisting
of the status code, headers, and a body. (More on why we chose this
format later).

Since the WebApplication will be handing a numeric status code back to
the WebServer, we'll also have to add a bit of logic to the WebServer to
know how to map this back to a "Reason Phrase" which the HTTP spec
expects us to include. For now we'll just focus on the status codes "200
Ok" and "404 Not Found". A complete list can be found in the HTTP
response spec.