Multi-threaded server
In order to process multiple requests at once, we're going to have to
introduce some threads to our web server. Concurrency & parallelism are
a very interesting and deep topic in Ruby, and we're not going to cover
them in depth here.

For our purposes, it'll be enough to create a new thread for each
connection that we receive from the server, process it in that thread,
and let the thread complete once the connection has been closed. This
allows the main thread to keep waiting for new connections using
@server.accept

The way that we _test_ this change is important too. I drove myself a
little crazy testing this out: I opened 5 different tabs in Google
Chrome, pointed them each at "http://127.0.0.1:8000/sleep/3" and then
refreshed all of them at once. What I expected to happen was that they
would all complete pretty much simultaneously. What _actually_ happened
was that one completed, then 3s later the next one completed, and so on
and so on. That's just as bad as our single threaded server!

Meanwhile I found that if I started one tab at `/sleep/3` and then in
another tab visited `/`, the second tab would still load before the
first! So clearly we _were_ getting some multithreading, but why didn't
it appear to be working for the first test?

It turns out that the HTTP _client_ we use (in this case Google Chrome)
also plays an important role. Modern browsers are incredibly complex,
and it seems that my browser was noticing that I was requesting the same
path in multiple tabs at the same and deciding to wait until the first
one finished to see if it needed to send the subsequent requests. If we
send slightly different requests, we can more easily see that we are
truly getting the concurrency that we're looking for.

With multithreading complete, lets turn back to the web application and
start expanding it to something that feels more familiar.